backend/requirements_update.txt
pandas==2.1.4
requests==2.31.0
supabase==2.28.3
python-dotenv==1.0.0